requirements.txt
coverage==4.2
Django==1.8
Pillow==2.3.0
psycopg2==2.5.4
pyflakes==1.2.3
python-jenkins==0.4.13
six==1.10.0
multi-key-dict==2.0.3